feat: release 9.0.0-alpha1
- refactor of PnC and perception
- support arm
- support plugins
- optimize config and project dir structure
- bugfix

+++ modules/tools/amodel/amodel/model_manage.py
@@ -23,11 +23,15 @@ import logging
 import os
 import requests
 import shutil
+import sys
+from pathlib import Path
 
 from amodel.model_meta import ModelMeta
 
 # APOLLO_ROOT_DIR
 WORKSPACE_PATH = os.getenv('APOLLO_ROOT_DIR', '/apollo')
+# Model install paths
+MODEL_INSTALL_PATH = "modules/perception/data/model"
 # MODEL_META_FILE_NAME
 MODEL_META_FILE_NAME = "apollo_deploy"
 
@@ -35,15 +39,8 @@ MODEL_META_FILE_NAME = "apollo_deploy"
 DOWNLOAD_TMP_DIR = "/tmp/"
 UNZIP_TMP_DIR = "/tmp/amodel/extract_path"
 
-# Model install paths
-MODEL_INSTALL_PATH = {
-  "3d_detection_lidar": "modules/perception/production/data/perception/lidar/models/detection",
-  "3d_segmentation_lidar": "modules/perception/production/data/perception/lidar/models/cnnseg",
-  "3d_detection_camera": "modules/perception/production/data/perception/camera/models/yolo_obstacle_detector",
-  "lane_detection_camera": "modules/perception/production/data/perception/camera/models/lane_detector",
-  "tl_detection_camera": "modules/perception/production/data/perception/camera/models/traffic_light_detection",
-  "tl_recognition_camera": "modules/perception/production/data/perception/camera/models/traffic_light_recognition",
-}
+# Download
+DOWNLOAD_REPOSITORY_URL = "https://apollo-pkg-beta.cdn.bcebos.com/perception_model"
 
 # Frame abbreviation
 FRAMEWORK_ABBREVIATION = {
@@ -102,10 +99,9 @@ def _get_all_model_metas():
       list: all installed models's meta
   """
   total_model_metas = []
-  for _, model_install_path in MODEL_INSTALL_PATH.items():
-    absolute_path = os.path.join(WORKSPACE_PATH, model_install_path)
-    model_metas = _get_model_metas(absolute_path)
-    total_model_metas.extend(model_metas)
+  absolute_path = os.path.join(WORKSPACE_PATH, MODEL_INSTALL_PATH)
+  model_metas = _get_model_metas(absolute_path)
+  total_model_metas.extend(model_metas)
   return total_model_metas
 
 def _display_model_list(model_metas, white_names=[]):
@@ -126,7 +122,7 @@ def _display_model_list(model_metas, white_names=[]):
   for index in range(len(model_metas)):
     model_meta = model_metas[index]
     if len(white_names) == 0 or (model_meta.name in white_names):
-      print("{:05}|{:<20}|{:<20}|{:<15}|{:<20}|{:%Y-%m-%d}".format(
+      print("{:<5}|{:<20}|{:<20}|{:<15}|{:<20}|{:%Y-%m-%d}".format(
         index,
         model_meta.name,
         model_meta.task_type,
@@ -155,6 +151,12 @@ def _is_url(url_str):
   """
   return url_str.startswith('https://') or url_str.startswith('http://')
 
+def _progress(prefix, cur, total):
+  bar_size = 50
+  cur_p = int(cur / total * bar_size)
+  print("{}[{}{}] {}/{}".format(prefix, "#"*cur_p, "."*(bar_size - cur_p),
+      cur, total), end='\r', file=sys.stdout, flush=True)
+
 def _download_from_url(url):
   """Download file from url
 
@@ -166,11 +168,22 @@ def _download_from_url(url):
   """
   local_filename = url.split('/')[-1]
   download_file = os.path.join(DOWNLOAD_TMP_DIR, local_filename)
+
+  # File is cached
+  if Path(download_file).is_file():
+    logging.warn(
+      "File downloaded before! use cached file in {}".format(download_file))
+    return download_file
+
   with requests.get(url, stream=True) as r:
     r.raise_for_status()
+    chunk_size = 8192
+    total_length = int(r.headers.get('content-length', 0)) // chunk_size
     with open(download_file, 'wb') as f:
-      for chunk in r.iter_content(chunk_size=8192):
+      for index, chunk in enumerate(r.iter_content(chunk_size)):
         f.write(chunk)
+        _progress("Downloading:", index, total_length)
+    print()
   return download_file
 
 def _unzip_file(file_path, extract_path):
@@ -207,16 +220,12 @@ def _get_install_path_by_meta(model_meta):
   Returns:
       str: model's install path
   """
-  perception_task = "{}_{}".format(model_meta.task_type, model_meta.sensor_type)
   file_name = "{}_{}".format(model_meta.name,
       FRAMEWORK_ABBREVIATION[model_meta.framework])
-  install_path = os.path.join(
-      WORKSPACE_PATH,
-      MODEL_INSTALL_PATH[perception_task],
-      file_name)
+  install_path = os.path.join(WORKSPACE_PATH, MODEL_INSTALL_PATH, file_name)
   return install_path
 
-def _install_model(model_meta, extract_path):
+def _install_model(model_meta, extract_path, skip_install):
   """Move model from extract_path to install_path
 
   Args:
@@ -229,7 +238,12 @@ def _install_model(model_meta, extract_path):
   install_path = _get_install_path_by_meta(model_meta)
   logging.debug("_install_model: {} -> {}".format(extract_path, install_path))
 
+  # Model exists
   if os.path.isdir(install_path):
+    if skip_install:
+      logging.warn(
+        "Skipped install, model already exist in {}.".format(install_path))
+      return True
     confirm = _user_confirmation(
       "Model already exists!!! Do you want to override {}? [y/n]:".format(
         model_meta.name))
@@ -242,7 +256,10 @@ def _install_model(model_meta, extract_path):
   print("Installed model in {}.".format(install_path))
   return True
 
-def amodel_install(model_path):
+def _construct_default_url(model_path):
+  return "{}/{}.zip".format(DOWNLOAD_REPOSITORY_URL, model_path)
+
+def amodel_install(model_path, skip_install = False):
   """amodel install command
 
   Args:
@@ -251,6 +268,10 @@ def amodel_install(model_path):
   if not model_path:
     logging.error("Input file is empty!!!")
     return
+
+  # If file does not exist and file is not url, use default url
+  if not os.path.isfile(model_path) and not _is_url(model_path):
+    model_path = _construct_default_url(model_path)
 
   # Download file
   if _is_url(model_path):
@@ -279,10 +300,8 @@ def amodel_install(model_path):
 
   # Install meta file
   extract_path = os.path.join(UNZIP_TMP_DIR, model_name)
-  is_success = _install_model(model_meta, extract_path)
-  if is_success:
-    print("Successed install {}.".format(model_meta.name))
-  else:
+  is_success = _install_model(model_meta, extract_path, skip_install)
+  if not is_success:
     logging.error("Failed install {}.".format(model_meta.name))
 
 
@@ -334,6 +353,10 @@ def amodel_remove(model_name):
       model_name (str): the model need to remove
   """
   # Find model index
+  if model_name is None:
+    logging.error("Input model name or number is empty!!!")
+    return
+
   total_model_metas = _get_all_model_metas()
 
   model_id = None
@@ -400,6 +423,10 @@ def amodel_info(model_name):
       model_name (str): model's name
   """
   # Find model index
+  if model_name is None:
+    logging.error("Input model name or number is empty!!!")
+    return
+
   total_model_metas = _get_all_model_metas()
 
   model_id = None
